Rentify/Backend/rentify/requirements.txt
asgiref==3.8.1
boto3==1.34.111
botocore==1.34.111
Django==5.0.6
django-cors-headers==4.3.1
django-storages==1.14.3
djangorestframework==3.15.1
jmespath==1.0.1
pillow==10.3.0
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
